--- tests/_utils/data/opensubtitles/opensub_pair_dev.post
@@ -1,100 +1,0 @@
-my name is josie .
-i put premium gasoline in her .
-josie , dont hang up
-about playing a part .
-ittotaily changed my swing , man .
-like , there goes another iemming .
-could you please teii me , what is the status of your story , ms. geiler ?
-i gotta talkto you .
-and i gotthem to iook atyourwriting , and , he agreed to meetwith you .
-josie, were gonna blow the iid off it .
-im 25 years old , and i have never reaily kissed a guy .
-theres stiii that one teacher who marches to her own drummer .
-love is a cancer , love aint benign
-you know that i need you
-the feds moved in . . . . . . and bingo , the bosses , they moved out .
-i was reflecting on your situation .
-. . . and get some answers for my friend .
-no , i don' t think you want
-you' re good , doc .
-what if you get tired of me ?
-that got away from me at the end there .
-no , for the big michigan state game .
-why do you have a girlfriend ?
-stress is a very powerful force .
-just hit the pillow .
-i could' ve saved him .
-i just let him die .
-i want to say i' m . . . .
-why don' t he just pop him ?
-that one' s mine .
-i think we' re going to get started here .
-some of you know me as joey boombots .
-elmer the fudd , tubby the tuba . . . . . . and once as miss phyilis levine .
-as mr. vitti' s . . . . . . consigliere . . .
-hands on top of the car !
-vitals stable and holding .
-that' s what i was talking .
-let' s have some music .
-your wife is very attractive , del .
-a mark , huh ?
-of course it matters !
-echoing i just know that there' s something really special out there for me .
-seems she was pretty familiar with old del on a regular basis , if you get my drift .
-oh , i know .
-no , i have to do this , or he' il die !
-you should thank that nurse for what she did .
-i can' t believe you' re doing all this for me , rosa .
-not the actor , whose real name is george , by the way , but his character , betty !
-i will shoot that bitch like she' d scratched my car !
-talk to him , huh ?
-i still owe him .
-remember the blind one last year ?
-you remember that line ?
-i love that line .
-that' s the same guy !
-listen , i' m sorry , i thought this would be the best way .
-call a doctor , bitch !
-she' s just tryin' to find herself .
-io , ti amo , david .
-first shot , you got it .
-what' s the difference ? !
-don' t do that .
-i haven' t got a bag .
-we ain' t started yet .
-opening night at the opera .
-you can only have two .
-that' s 1 . 70 , that' s it .
-no , i think .
-you know , the guys in the back tunnelling under the nail shop ?
-but it' s great .
-what are we doin' in a dress shop ?
-what are your duties as chairman ?
-i want a lot of nuts and crudites .
-nothing being served requires them .
-excuse me , i have to go find my wife .
-but i have come into possession of an incredible damon dexter that someone' s trying to sell , is he . ?
-but i' m not going to museums .
-you should be able to cultivate the right kinds of friends .
-she' s looking for you .
-it' s a tough safe , i' m out of practice .
-get out there and stand watch .
-i' m sorry i blew our cookie company but i' m even more sorry i blew our marriage .
-as usual , i come up empty .
-it beats your former residence .
-two years in the joint i waited for you .
-a little hole in the wall .
-i don' t know what' s gotten into you .
-you' re benny from cell block eight !
-no , it was sarcastic .
-it wasn' t sarcastic .
-we ain' t started yet .
-you' re gonna be fine .
-you got a mirror ?
-where' s the sugar ?
-yes , it' s true , we were gonna rob the bank .
-this is pretty much where i operate .
-it think it' s the renaissance or the magna carta or something .
-then there' s ex police officer ken deloach , vice president in charge of distribution .
-first thing i did was make sure the building was fireproof .
-truffles , by all means .
