--- tests/_utils/data/opensubtitles/opensub_pair_dev.response
@@ -1,100 +1,0 @@
-imfrom scranton , pennsylvania , the keystone state .
-josie , i told you .
-they do it to aii the new kids .
-and thats the theme of as you like it .
-ah , no problem , man .
-i cant believe that i could be happy
-oh , uh , great .
-i need some air .
-you believe in me that much ?
-there is no iid .
-a geekto the core , most of my childhood years were spent . doing extra homework i requested from the teacher .
-those girls are stiii there , the ones that , even as you grow up , wiii remain the most beautifui girls . you have ever seen close up .
-love is what keeps you going aii the time now
-never knew what love was
-along comes a farmer .
-it' s interesting what you were saying .
-it could be interpreted wrong .
-my roster of patients is full .
-i' il be getting in touch with you .
-. . . and completely lose interest in each other .
-i couldn' t get it up last night .
-of course , sexually !
-you' re moralizing with me now ?
-i doubt there' s anything physically wrong with you .
-see how you feel .
-but you were mad at him .
-i let him die !
-dad , i' m sorry .
-it' s embarrassing , standing here like this .
-you get the' 72 chevy . . .
-we got a lot to talk about here today , and . . .
-. . . and to welcome some of the new ones from all over the country .
-but that was at a party , years ago .
-badda bing badda boom betty boop .
-take a pill , robocop !
-close her up , will you ?
-yeah , i was talking about .
-swinging sax you guys like music ?
-have i seen her before ?
-it' s ok , david .
-now , i want to know the particulars of your wrongdoing .
-very unprofessional , wesley !
-and half the other guys in this town , including yourself , i believe .
-and i' m real sorry about all this .
-it' s ok , i' ve seen it done once .
-iris , who was that remarkable nurse ?
-i just wanna make sure that you get your . fairy tale ending , or whatever .
-are you having a nervous breakdown ?
-do not talk like that .
-you came 1, 500 miles for this .
-he never said anything .
-and he got the emmy thing for that ?
-i said that to leslie during her funeral .
-well . . you said it to her . . but you meant it for me , didn' t you ?
-no , no , no .
-throw you into it , y' know ?
-what the hell is that ?
-you' re my only friend .
-ti amo , geri .
-is that why tommy beale' s coming over , and denny ?
-tell fortunes , you look like a gypsy , make it a travel agency .
-i gotta talk to her .
-you' il take' em like this ?
-what are you waitin' for , the drilling season ?
-i love you , frenchy , you know why ?
-what are you stopping for ?
-thank you very much .
-i think we could do it .
-what are you talking about ?
-and we' re expanding , there are men in the back tunnelling .
-at least it' s a sunday , we can patch this up and get out .
-the board meets twice a week and we take up issues .
-what did we say for openers ?
-you never know whose fingers need washing .
-i saw a beautiful painting of fruit at the met or the whitney , it might have been the holocaust .
-i' m not aware of him .
-the . pictures spook me out .
-anyway , i , you know .
-now i' ve closed the safe .
-did they see you come up here ?
-i' m gonna do this again .
-what are you gonna do ?
-we could pawn this .
-think i like spending our anniversary staring at you through glass , talking over a phone ?
-and what did i get for it ?
-and it' s for rent , frenchy .
-did we not have dreams when i first met you ?
-oh , god , ray !
-there was nothing sarcastic about it !
-oh , my god !
-what do you mean ?
-this time , a month or two , we' il be in miami .
-i don' t got a mirror .
-frenchy wants more sugar .
-but we screwed up everything , except for the cookie shop .
-frenchy decorated this place .
-but that' s where it' s from .
-and benjamin borkowski , vice president in charge of plant safety .
-and lest they be accused of being a mostly male club ,
-but i hate those thin pieces .
