--- tests/_utils/data/opensubtitles/opensub_pair_test.post
@@ -1,100 +1,0 @@
-if it were anyone but turold' s son .
-in the fortress , you will face more than the slayers .
-in a cave on the highest peak .
-i might have been sucked to my death while you stood there gawking !
-if the invaders conquer , you' il be slaves with the rest of us .
-you know what i mean ?
-how did you know i had sons ?
-i saw you save the boy from a spear .
-i recognise the place now .
-mamma' s boy is holding out .
-now i see why you picked this camp .
-it just wasn' t there anymore .
-if he keeps stuffing his face like that , soon there won' t be any rocks here either .
-we are on the same mission .
-so , she cannot save us .
-if i don' t save her , she' il die .
-you and your scientific work .
-brave men discover they are really cowards .
-we will not harm you .
-i was the one chosen to stop the nothing .
-one grain of sand .
-oh , and harker said i' m supposed to collect all your files on tom bishop .
-it' s all i could find .
-due to the nature of this task force , nathan , there will be certain information which you don' t need to know .
-do you trust him ?
-he knew about bishop before he got here .
-i talked to his neighbors , his teachers , his mom scoutmaster
-and then i watched . and waited .
-look him straight in the eye , and ask him to go back . to the place he just fled and spy for us .
-if they stop us , don' t say a word .
-your scotch is older than she is .
-they don' t know .
-this is a whole other game .
-but tonight , they saved your life .
-sideshow was a bugging op .
-bishop wasn' t on sideshow .
-this is frank nall , live in hong kong .
-is it relevant to the issue at hand ?
-that sounds iike fun ?
-that makes you feei better ?
-oh , and you look out for spriggen and banshee , my girl .
-with it , my empire is eternal .
-oh , it is .
-your fine sensibilities have left us here to rot !
-just the two of us .
-you killed her partner , didn' t you ?
-grandfather , get up .
-defend it from who ?
-the essence of mortal kombat . is not about death . but life .
-come to a little tournament , " he said .
-like my daddy taught me .
-why should i worry about her ?
-you' re very foolish .
-those were $500 sunglasses , asshole .
-i would not have helped you .
-i forgive you for letting me die .
-i' il take care of it when i get back .
-he said he was your husband' s doctor .
-what do you expect ?
-the idea that a simple man could develop the ability to tune
-can anybody hear me ? !
-you can' t get out of the city .
-the one they' ve been looking for .
-you burnt your arm pretty bad in the fire .
-how can it be night already ?
-i understand you now , mr murdoch .
-this . affair of yours , whatever it is , the thing you' re supposed to have done , you didn' t do it .
-lnspector , he is more disturbed than we thought .
-you have had dozens of lives before now .
-they' re dying , you see ?
-but he can also lead us to what we seek .
-you will survive , john .
-i' m dying , john .
-am i trying to be too perfect ?
-so , if i needed a doctor , you could be it .
-who is this guy ?
-what about the mayor ?
-when i introduce you , if i say who you are ,
-yeah , he is .
-what are you doing here ?
-will that hold you , joe ?
-but , well , ten hours later , i feel like a fool .
-you were hatching ideas , making plans , arriving at decisions , right ?
-and call my family .
-and perhaps cheating on your french philosophers exam at the groton school . was an expedient way to get your diploma , and perhaps it wasn' t .
-i' m tired of his fancy name and his fancy offer .
-now , look , i know you' re down .
-that' s what he said .
-i am looking for baltus van tassel .
-taken by the headless horseman .
-i was young masbath , but now the only one .
-i am your man , sir .
-bring the widow in .
-are you all right ?
-i will pit myself against a murdering ghost .
-wait , he' s not after you .
-katrina , i tried to stop brom .
-steenwyck knew the secret .
-up pops a widow with a claim on the fortune .
-but murder begets murder .
