--- tests/_utils/data/opensubtitles/opensub_pair_test.response
@@ -1,100 +1,0 @@
-colwyn is a great fighter .
-you will face the beast , who is their leader .
-without the glaive , you will never be able to reach lyssa .
-where is this place ?
-noble sentiments , but we fight for profit .
-the profit' s freedom .
-if the slayers conquer krull , your sons will be enslaved forever .
-that' s what friends are supposed to do for each other .
-we' re in sight of the trees , brother .
-if you ain' t got money , you go in the garbage .
-a delicious looking limestone rock !
-nothing was there anymore .
-night hob , this could be serious !
-after all , if the empress can' t save us . . who could ?
-but there just might be one chance .
-there' s a terrible nothing sweeping over the land .
-what the boy needs now is one of my potions .
-confronted with their true selves . . most men run away screaming .
-we have been waiting for you a long time .
-but i lost the auryn .
-it is all that remains of my vast empire .
-finally get to use one of these burn bags .
-packed , misplaced you know how it is .
-a week ago , tom bishop disappeared in hong kong .
-he' s a man who got the job done .
-and he played dumb .
-pulled his dad' s records from the pentagon .
-this consignment is a complete misdirect .
-he got so good at it he could close in a single afternoon , send the guy out with a smile .
-remember , you' re drunk .
-am i supposed to feel bad about that ?
-they don' t fuckin' know .
-and it' s serious and it' s dangerous and it' s not one you want to lose .
-can we have the relevant files on ann cathcart ?
-listening to government offices in eastern china .
-he wasn' t working for us .
-guess that limits our options .
-i know who bishop was after .
-yeah . that' s what i think
-if that' s what you want , i' ii go with you
-these woods be thick with them .
-before i ruled only the night .
-i' m warm and alive . and want to be in your arms .
-i' m behind you all the way , gump .
-some . simple conversation .
-she' il follow you into hell .
-this isn' t your god of thunder .
-your world is but one of many realms .
-mortal men and women . defending your own world .
-be good for the career , " he said .
-but what i just saw wasn' t very fair .
-princess kitana is 10, 000 years old . and the rightful heir to the throne of outworld .
-the true sign of a hero .
-this is where you fall down .
-ln the black tower you will face three challenges .
-it wasn' t my fault .
-see that you do .
-he wanted to speak with you .
-weren' t you looking for the human soul ?
-ls absurd , i know , but what other explanation is there ?
-let the tuning commence .
-believe me , i' ve tried .
-you' re not a killer .
-what does it mean , johnny ?
-what happened to the day ?
-i remember that which you do not , what you' ve been missing .
-i don' t believe it ever happened .
-i may not be the judge of personality that you are , doctor , but you' re the one who looks disturbed to me .
-you just happened to wake up while i was imprinting you with this one .
-their entire race is on the brink of extinction .
-what the doctor calls the soul .
-you will find strength within yourself , and you will prevail .
-your imprint is not agreeable with my kind .
-oh , but i want it to be so exquisite .
-i could be her .
-tell me who you are .
-he said he' s going to be there with bells on .
-i don' t think anyone will stay for dinner .
-shall we sit down ?
-you know each other ?
-incidentally , joe what ?
-i just don' t get it .
-yeah , i guess so .
-i' d like them to have dinner with me tonight .
-be that as it may , drew , a question can often be argued both ways .
-i' m not going for it .
-but you know , when you' re down , there' s no place to go but up .
-that' s very interesting .
-i am his daughter , katrina .
-taken back to hell .
-masbath at your service .
-and a brave one , too .
-this is most irregular , constable .
-yes , thank you .
-who' s with me ?
-i' il get him .
-drink this down , it will make you sleep .
-he performed the marriage .
-off with her head .
-it was the servant . .
