--- tests/_utils/data/opensubtitles/opensub_pair_train.post
@@ -1,100 +1,0 @@
-catch afailing star and put it in your pocket .
-you know how in some movies . they have a dream sequence , only they dontteii you its a dream ?
-this is so not a dream .
-itwasnt supposed to be iike this .
-i was justtrying to do my job , and then things happened .
-weil , iife happened .
-and now im here .
-im notthe kind of girl . who does things iike this .
-love may come and tapyou on the shoulder .
-i was the youngest copy editor atthe chicago sun times .
-standard british cailsfor " re .
-so unless youre a pompous american ,
-i would go with " er .
-i had my own office , unlimited stationary supplies , and a professional , dedicated assistant .
-thought of this awesome new phrase . " power is powerful .
-excuse me , merkin .
-um , i noticed thatthere werent any index cards .
-i cant do notes withoutthem .
-hes workin , ok ?
-take thatto the bank .
-im just a iittle tense .
-i mean, this whole office is not feng shui .
-aii the desks arefacing , iike , evil, you know ?
-need it by 5 : 00 .
-hopefuily the copys not a mess .
-its " hoped " that its not a mess . " hopefuily " is an adverb .
-it means " with hope .
-you and roger in op/ ed .
-oh , man, who told you ?
-dont make me send you another memo . about my policy on interoffice dating .
-its " intraoffice dating .
-and what exactly is your policy, augustus ?
-did you getthose story ideas that i ieft on your desk ?
-morning , mr. strauss .
-geiler, weve been through this .
-youre a great copy editor .
-youre maybe my best copy editor .
-youre not a reporter .
-he grabs the buii by the bails .
-you dontthink i can grab a buils bails ?
-theyre very messy , and youre aii about order and control , and getting me my copy by 5 : 00 .
-i can be out of control .
-copy by 5 : 00 !
-josie , maybe gus has a point .
-you know , its not gonna kiii you to relax and have somefun .
-hey, you know , rogers got thisfriend marshaii in editing .
-teii me, when is the iasttime thatyou went on a reai date ?
-im concentrating on my career right now .
-the right guy , hes outthere .
-im just not gonna go kiss a whole bunch of iosers . to getto him .
-yeah , butyou know what ?
-if youve never kissed a guy before , we have bigger problems than the underwear .
-you are a writer .
-what do you guys think ?
-thats exactly what i was thinking .
-to save me from ruin
-when i say that im ok
-surely youre not happy now
-you no ionger play the game
-the meeting just began .
-let me start by saying i was very impressed . with that investigative piece that dutton did . on pesticides in our supermarkets .
-but since the trib did a piece on the same subject only better , yourefired .
-bye bye , dutton .
-heilo , everybody else stiii working here .
-lets celebrate by doing another undercoverfeature .
-now, you know that my inspirationforthese stories . come , of course , from my personai iife .
-i mean, who wouldve thought that myforay into hair plugs . wouldve resulted in iast months cover story ? " hairtoday , gone tomorrow . " a great story .
-so anyway , iast night ,
-and my younger boy starts to choke .
-and i realized that my son was ailergic to peanuts .
-i said to myself , " holy shit !
-i dont even know my own kids .
-how much do we know about these high schooi kids ?
-what are they thinking ?
-myfirst undercoverfeature by josie geiler !
-this is a iot of pressurefor yourfirst piece, thats ail .
-i mean, this is not a half a page article .
-this is a major undercover piece , ok ?
-im gonna straighten this aii out .
-neither of you guys think i can do this ?
-thats notwhat were saying, josie .
-so this is my chance .
-dont make me give you my memo on interoffice hugging , ok ?
-now comes the hard part .
-oh , my gosh .
-you getto be 17 again .
-why dontyou just borrow moms car ?
-i cant use a minivan .
-its justfor a couple of months .
-a couple of months is iike 10 years in bambi iife .
-its just my sister .
-aloha , my ass .
-shake it easy , buddy .
-butyou could start before next semester .
-you could even talkto her about a basebaii scholarship .
-its too iate, josie , you know ?
-im not going to coilege .
-im not playing basebaii anymore .
-this is my iife .
-this is a iuau that seils packing material .
