--- tests/_utils/data/opensubtitles/opensub_pair_train.response
@@ -1,100 +1,0 @@
-you know how in some movies . they have a dream sequence , only they dontteii you its a dream ?
-this is so not a dream .
-itwasnt supposed to be iike this .
-i was justtrying to do my job , and then things happened .
-weil , iife happened .
-and now im here .
-catch afailing star and put it in your pocket .
-i mean, two months ago , you couldnt have picked me out of a crowd .
-i was the youngest copy editor atthe chicago sun times .
-standard british cailsfor " re .
-so unless youre a pompous american ,
-i would go with " er .
-i had my own office , unlimited stationary supplies , and a professional , dedicated assistant .
-thought of this awesome new phrase . " power is powerful .
-totai t shirt , right ?
-um , i noticed thatthere werent any index cards .
-i cant do notes withoutthem .
-hey , hey , hey .
-take thatto the bank .
-its power , baby .
-i mean, this whole office is not feng shui .
-aii the desks arefacing , iike , evil, you know ?
-uh , do you think i should bleach my jeans .
-hopefuily the copys not a mess .
-its " hoped " that its not a mess . " hopefuily " is an adverb .
-it means " with hope .
-you have it defining the copy , and im pretty sure the copy doesnt havefeelings .
-oh , man, who told you ?
-roger in op/ ed .
-its " intraoffice dating .
-and theyre not dating .
-that if youre not getting any , then no one can ?
-morning , mr. strauss .
-yes, the blind foster home mother .
-youre a great copy editor .
-youre maybe my best copy editor .
-youre not a reporter .
-youve used 5 of my ideas .
-you dontthink i can grab a buils bails ?
-geiler, you dontwant a reporters iife .
-i can be out of control .
-copy by 5 : 00 !
-josie , maybe gus has a point .
-you know , its not gonna kiii you to relax and have somefun .
-hey, you know , rogers got thisfriend marshaii in editing .
-maybe we can aii go on a double date or something .
-im concentrating on my career right now .
-do you own any colored underwear ?
-im just not gonna go kiss a whole bunch of iosers . to getto him .
-yeah , butyou know what ?
-sometimes kissing iosers can be a reailyfun diversion .
-ive kissed a guy .
-people say im crazy .
-where do you think it should go ?
-to save me from ruin
-when i say that im ok
-weil, they iook at me kind of strange
-you no ionger play the game
-the meeting just began .
-let me start by saying i was very impressed . with that investigative piece that dutton did . on pesticides in our supermarkets .
-but since the trib did a piece on the same subject only better , yourefired .
-dutton , did you hear me ?
-heilo , everybody else stiii working here .
-lets celebrate by doing another undercoverfeature .
-now, you know that my inspirationforthese stories . come , of course , from my personai iife .
-i mean, who wouldve thought that myforay into hair plugs . wouldve resulted in iast months cover story ? " hairtoday , gone tomorrow . " a great story .
-so anyway , iast night ,
-im sitting home with myfamily having dinner .
-and i realized that my son was ailergic to peanuts .
-i said to myself , " holy shit !
-i dont even know my own kids .
-how much do we know about these high schooi kids ?
-what are they thinking ?
-and then it hit me .
-josie , maybe you should turn it down .
-i mean, this is not a half a page article .
-this is a major undercover piece , ok ?
-lookwhat rigfortjust did to dutton in there , and that guys his cousin .
-neither of you guys think i can do this ?
-thats notwhat were saying, josie .
-anita, when you wanted to seduce the guy in the mailroom , and you thoughtyou couldnt iearn spanishfast enough , who quizzed you on yourverbs ?
-weil , im not holding yourjobforyou , geiler .
-now comes the hard part .
-oh , my gosh .
-you getto be 17 again .
-ok, whats thefirst thing youre gonna need ?
-i cant use a minivan .
-i need a cheap car .
-a couple of months is iike 10 years in bambi iife .
-its so weird thatyou name your cars .
-aloha , my ass .
-shake it easy , buddy .
-hey , rob , have you talked to the admissions iady . from lakeshore community coilege yet ?
-you could even talkto her about a basebaii scholarship .
-its too iate, josie , you know ?
-im not going to coilege .
-im not playing basebaii anymore .
-this is my iife .
-this is a iuau that seils packing material .
-you had such a shot at playing coilege bail .
