+++ tests/file_utils/data/glove.6B.50d.zip
