+++ tests/file_utils/data/mscoco.zip
