+++ tests/file_utils/data/mscoco/mscoco_dev.txt
