+++ tests/file_utils/data/mscoco/mscoco_test.txt
