+++ tests/file_utils/data/mscoco/mscoco_train.txt
