+++ tests/file_utils/dummy_coai/MSCOCO.json
