+++ tests/file_utils/test_file_utils.py
@@ -8,7 +8,7 @@ import json
 import shutil
 from checksumdir import dirhash
 
-from cotk._utils.file_utils import get_resource_file_path, import_local_resources
+from cotk.file_utils import get_resource_file_path, import_local_resources
 
 @pytest.fixture
 def r_mock():
@@ -19,8 +19,8 @@ class TestFileUtils():
 	def test_get_resource(self, r_mock):
 		r_mock.get('http://coai.cs.tsinghua.edu.cn/', text='coai')
 
-		cache_dir = str(pathlib.Path('./tests/_utils/dataset_cache'))
-		config_dir = str(pathlib.Path('./tests/_utils/dummy_coai'))
+		cache_dir = str(pathlib.Path('./tests/file_utils/dataset_cache'))
+		config_dir = str(pathlib.Path('./tests/file_utils/dummy_coai'))
 
 		with pytest.raises(FileNotFoundError) as excinfo:
 			get_resource_file_path('resources://coai', cache_dir=cache_dir, config_dir='wrongpath')
@@ -35,7 +35,7 @@ class TestFileUtils():
 		assert "source wronglink wrong" in str(excinfo.value)
 
 		res_path = get_resource_file_path('resources://coai', cache_dir=cache_dir, config_dir=config_dir)
-		
+
 		assert res_path == os.path.join(cache_dir, '146ce545f2ed0a8767aadae8f2921f7951df817b39b8f7d0db48bce87e3eaf69')
 		assert os.path.exists(res_path)
 
@@ -50,12 +50,12 @@ class TestFileUtils():
 		with open(meta_path, 'r', encoding='utf-8') as meta_file:
 			meta = json.load(meta_file)
 			assert meta['local_path'] == res_path
-		
+
 		shutil.rmtree(cache_dir)
 
 	def test_download_resource(self):
-		cache_dir = str(pathlib.Path('./tests/_utils/dataset_cache'))
-		config_dir = str(pathlib.Path('./tests/_utils/dummy_coai'))
+		cache_dir = str(pathlib.Path('./tests/file_utils/dataset_cache'))
+		config_dir = str(pathlib.Path('./tests/file_utils/dummy_coai'))
 		res_path = get_resource_file_path('resources://test@amazon', cache_dir=cache_dir, config_dir=config_dir)
 		res_path = get_resource_file_path('resources://test', cache_dir=cache_dir, config_dir=config_dir)
 
@@ -67,12 +67,12 @@ class TestFileUtils():
 		with open(meta_path, 'r', encoding='utf-8') as meta_file:
 			meta = json.load(meta_file)
 			assert meta['local_path'] == res_path
-		
+
 		shutil.rmtree(cache_dir)
 
 	def test_download_data(self):
-		cache_dir = str(pathlib.Path('./tests/_utils/dataset_cache'))
-		config_dir = str(pathlib.Path('./tests/_utils/dummy_coai'))
+		cache_dir = str(pathlib.Path('./tests/file_utils/dataset_cache'))
+		config_dir = str(pathlib.Path('./tests/file_utils/dummy_coai'))
 		res_path = get_resource_file_path('https://cotk-data.s3-ap-northeast-1.amazonaws.com/test.zip', cache_dir=cache_dir)
 		res_path = get_resource_file_path('https://cotk-data.s3-ap-northeast-1.amazonaws.com/test.zip', cache_dir=cache_dir)
 
@@ -84,13 +84,13 @@ class TestFileUtils():
 		with open(meta_path, 'r', encoding='utf-8') as meta_file:
 			meta = json.load(meta_file)
 			assert meta['local_path'] == res_path
-		
+
 		shutil.rmtree(cache_dir)
 
 	def test_import_local_resource(self):
-		cache_dir = str(pathlib.Path('./tests/_utils/dataset_cache'))
-		config_dir = str(pathlib.Path('./tests/_utils/dummy_coai'))
-		data_dir = str(pathlib.Path('./tests/_utils/data'))
+		cache_dir = str(pathlib.Path('./tests/file_utils/dataset_cache'))
+		config_dir = str(pathlib.Path('./tests/file_utils/dummy_coai'))
+		data_dir = str(pathlib.Path('./tests/file_utils/data'))
 
 		with pytest.raises(ValueError) as excinfo:
 			import_local_resources('test', local_path=os.path.join(data_dir, 'test.zip'), cache_dir=cache_dir, config_dir=config_dir)
@@ -104,7 +104,7 @@ class TestFileUtils():
 
 		with pytest.raises(ValueError) as excinfo:
 			import_local_resources('resources://test', local_path=os.path.join(data_dir, 'test.zip'), cache_dir=cache_dir, config_dir=config_dir, ignore_exist_error=True)
-			import_local_resources('resources://test', local_path=os.path.join(data_dir, 'test.zip'), cache_dir=cache_dir, config_dir=config_dir, ignore_exist_error=False)		
+			import_local_resources('resources://test', local_path=os.path.join(data_dir, 'test.zip'), cache_dir=cache_dir, config_dir=config_dir, ignore_exist_error=False)
 		assert "resources existed. If you want to delete the existing resources." in str(excinfo.value)
 
 		assert res_path == os.path.join(cache_dir, '9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08')
@@ -115,5 +115,5 @@ class TestFileUtils():
 		with open(meta_path, 'r', encoding='utf-8') as meta_file:
 			meta = json.load(meta_file)
 			assert meta['local_path'] == res_path
-		
+
 		shutil.rmtree(cache_dir)
